Remove the `strategy` package (#79)
* Remove ExtrapolatingStrategy and its unit tests
* Remove IsolatingStrategy and its unit tests
* Remove PartitioningStrategy and its unit tests
* Rename aggregation.py to aggregate.py
* Move UnifyingStrategy, Reshape and Matrixify to aggregate.py
* Remove the strategy package in sources and unit tests
* Make UnifyingStrategy, Reshape and Matrixify protected
* Rename _UnifyingStrategy to _AggregateMatrices
* Replace function make_aggregation by transform Aggregate
* Remove property testers for strategies: now directly test _AggregateMatrices with functions
* Improve the randomness of jacobian matrix dicts generation in tests by using a random generator
* Fix mistake in test_disunite_wrong_vector_length
* Add docstrings to testing functions in test_aggregate.py when missing
---------
