--- src/torchjd/autojac/_transform/strategy/__init__.py
@@ -1,4 +1,0 @@
-from .extrapolating import ExtrapolatingStrategy
-from .isolating import IsolatingStrategy
-from .partitioning import PartitioningStrategy
-from .unifying import UnifyingStrategy
