--- tests/unit/autojac/_transform/strategy/test_isolating.py
@@ -1,15 +1,0 @@
-import pytest
-
-from torchjd.autojac._transform.strategy import IsolatingStrategy
-
-from .utils import EmptyDictProperty, ExpectedStructureProperty, aggregator, keys
-
-
-@pytest.mark.parametrize("strategy", [IsolatingStrategy(aggregator, keys)])
-class TestIsolatingStructure(ExpectedStructureProperty):
-    pass
-
-
-@pytest.mark.parametrize("strategy", [IsolatingStrategy(aggregator, [])])
-class TestIsolatingEmpty(EmptyDictProperty):
-    pass
