--- tests/unit/autojac/_transform/strategy/utils/__init__.py
@@ -1,2 +1,0 @@
-from .inputs import aggregator, jacobian_matrix_dicts, keys
-from .property_testers import EmptyDictProperty, ExpectedStructureProperty
